requirements.txt
tensorflow== 2.17.0
numpy
pandas
matplotlib
opencv-python
flask
pillow
scikit-learn
